Add boolean flag for automatic responses to banners to the CLI.
Some servers require the user to respond to a so-called banner upon login. Not providing an answer results in the login attempt failing. To facilitate automating logging into such servers the --autorespond flag will automatically respond 'yes' to a login banner. This flag defaults to false and supports the complementary --noautorespond flagshould the default ever change.